requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
# RAG dependencies (ensure wheels available for Python 3.13)
numpy==2.1.3
sentence-transformers==2.2.2
python-multipart==0.0.6
pdfminer.six==20221105
httpx==0.27.0
